test: add more mutation cases
